+++ app/src/main/java/br/com/connectattoo/ui/registration/TattooClientRegistrationFragment.kt
@@ -2,7 +2,6 @@ package br.com.connectattoo.ui.registration
 
 import android.graphics.Color
 import android.os.Build
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -155,36 +154,27 @@ class TattooClientRegistrationFragment : UserRegistration<FragmentTattooClientRe
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
-    override fun validateUserRegistration(action: Int) {
-        val userRegisterData = ClientData(
+    private fun createRegisterUserData() : ClientData {
+        return ClientData(
             name = this.name.text.toString(),
             email = this.email.text.toString(),
             birthDate = formatBirthDate(this.birthDay.text.toString()),
             password = this.password.text.toString(),
             termsAccepted = this.terms.isChecked
         )
+    }
+    @RequiresApi(Build.VERSION_CODES.O)
+    override fun validateUserRegistration(action: Int) {
+        val userRegisterData = createRegisterUserData()
+
         apiService.registerUser(userRegisterData).enqueue(object : Callback<TokenData> {
             override fun onResponse(call: Call<TokenData>, response: Response<TokenData>) {
-
-                if (response.isSuccessful) {
-                    val responseBody = response.body()
-                    Log.d("Response", "Retorno: $responseBody")
-                    if (responseBody != null) {
-                        token = responseBody.accessToken
-                        Log.d("Token", "Token: $token")
-                        response.body()?.let {
-                            findNavController().navigate(action)
-                        }
-                    }
-                } else {
-                    when (response.code()){
-                        404 -> showValidationError("A URL de destino não foi encontrada.")
-                        409 -> showValidationError("Email já cadastrado!!!")
-                        else -> showValidationError("Erro: ${response.code()}")
+                registrationResponse(action, response) {
+                    response.body()?.let {
+                        findNavController().navigate(action)
                     }
                 }
             }
-
             override fun onFailure(call: Call<TokenData>, t: Throwable) {
                 showValidationError("Falha na conexão com a internet")
             }
